Make EngineNet return network statistics + few bug fixes
Update EngineNet behavior to return the netsize and average degree of lazily initialized predefined network. Add option for degree-dependent node size in drawing. If args.update_after is an integer, correct the edge update entry to match the closest day to the current time. Add new charts for SocialEvol experiment.